requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
certifi
yt-dlp==2023.12.30
python-magic==0.4.27